requirements.txt
# NIZAM COP — Python dependencies
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database (PostgreSQL + TimescaleDB)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3

# Auth (JWT + password hashing)
# PyJWT (maintained, timezone-aware) replaced python-jose which used utcnow()
# Upper bound pinned so a PyJWT 3.x breaking change can't silently land at install time.
PyJWT>=2.9.0,<3.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP client
httpx==0.27.2

# Image processing (EO agent) — headless variant saves ~20MB + no GUI deps
opencv-python-headless==4.10.0.84

# Config
python-dotenv==1.0.1

# AI / ML
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
# torch installed separately in Dockerfile via CPU-only wheel index
# (default torch wheel pulls ~800MB of CUDA libs we don't use)

# MQTT (mqtt_adapter)
paho-mqtt>=2.0.0

# OpenTelemetry (optional — gracefully disabled if not installed)
# Upper bounds keep the three OTEL packages on the same 1.x minor release;
# skew across them causes subtle ABI breaks.
opentelemetry-sdk>=1.25.0,<2.0.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.46b0,<1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# ── NIZAM Counter-UAS Servisleri (Faz 1+) ────────────────────────

# Faz 1 — Kamera tespiti
ultralytics>=8.3.0,<9.0.0

# Messaging (Faz 1, 2, 3, 4 için gerekli)
nats-py>=2.6.0,<3.0.0

# Faz 3 — Füzyon
filterpy>=1.4.5,<2.0.0
scipy>=1.11.0

# Faz 5 — CoT/TAK
pytak>=6.0.0
lxml>=4.9.0

# Faz 6 — Karar (YAML ROE)
pyyaml>=6.0.0

# Prometheus
prometheus-client>=0.19.0

# Optional LLM stack (Faz 6 LangGraph advisor) — install with:
#   pip install "langgraph>=0.2.0" "anthropic>=0.40.0" "llama-index>=0.12.0"
# Şimdilik requirements dışı — NIZAM_DECISION_LLM_ENABLED=true iken gerekli.

# Faz 7 — MAVSDK (opsiyonel, yoksa mock mode)
# mavsdk>=2.0.0  # sadece PX4 SITL / gerçek donanım varsa
